Moved the file list/zip creation outside of the builder class.
